Allow for non-integer subscription quantity matches...

+++ src/main/java/com/suse/matcher/facts/Match.java
@@ -41,7 +41,7 @@ public class Match implements Comparable<Match> {
     public Long subscriptionId;
 
     /** The number of subscriptions used in this match. */
-    public Integer quantity;
+    public Double quantity;
 
     /** The kind. */
     public Kind kind;
@@ -55,7 +55,7 @@ public class Match implements Comparable<Match> {
      * @param quantityIn the number of subscriptions used in this match
      * @param kindIn the match kind
      */
-    public Match(Long systemIdIn, Long productIdIn, Long subscriptionIdIn, Integer quantityIn, Kind kindIn) {
+    public Match(Long systemIdIn, Long productIdIn, Long subscriptionIdIn, Double quantityIn, Kind kindIn) {
         systemId = systemIdIn;
         productId = productIdIn;
         subscriptionId = subscriptionIdIn;
@@ -95,7 +95,7 @@ public class Match implements Comparable<Match> {
      *
      * @return the quantity
      */
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
